Renamed Interpreter to ApacheInterpreter and moved platform specific code back to the jvm package.
