+++ core/jvm/src/main/scala/hammock/hi/platformspecific.scala
